+++ src/main/java/duke/DukeException.java
@@ -1,5 +1,9 @@
 package duke;
 
+/***
+ * The DukeException extends Exception containing the message to throw when faces exception
+ * The subclasses contains different messages for different exceptions when thrown
+ */
 public class DukeException extends Exception {
     public String message;
 
@@ -7,3 +11,16 @@ public class DukeException extends Exception {
         this.message = message;
     }
 }
+
+class TaskNotExist extends DukeException {
+    public TaskNotExist() {
+        super("Invalid value, there isn't this much tasks in the list :-( ");
+    }
+}
+
+class MissingDescription extends DukeException {
+    public MissingDescription() {
+        super("OOPS!!! The description of a task cannot be empty");
+    }
+}
+
