--- src/main/java/duke/MissingDescription.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class MissingDescription extends DukeException {
-    public MissingDescription() {
-        super("OOPS!!! The description of a task cannot be empty");
-    }
-}
