+++ src/main/java/duke/Storage.java
@@ -8,9 +8,17 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 
+/***
+ * The class to store and load the ArrayList contents into a txt file and vice-versa
+ */
 public class Storage {
     private final File file;
 
+    /***
+     *
+     * @param filePath the place where the txt file is stored
+     * @throws IOException If the file cannot be found
+     */
     public Storage(String filePath) throws IOException {
         this.file = new File(filePath);
         if (!file.exists()) {
@@ -19,6 +27,10 @@ public class Storage {
         }
     }
 
+    /***
+     *  Save the data from the arraylist into the txt file created/stored
+     * @param tasks an arraylist of all the tasks
+     */
     public void saveData(ArrayList<Task> tasks) {
         try {
             File userData = new File("userData");
@@ -44,6 +56,10 @@ public class Storage {
         }
     }
 
+    /***
+     * A method to laod the data from the txt file to the arraylist
+     * @param tasks an arraylist of all the tasks
+     */
     public void loadData(ArrayList<Task> tasks) {
         try {
             File userData = new File("userData");
@@ -76,23 +92,11 @@ public class Storage {
         }
     }
 
-    public void realTimeSave(ArrayList<Task> tasks) {
-        try {
-            File dukeTxt = new File("duke.txt");
-            if (!dukeTxt.exists()) {
-                dukeTxt.createNewFile();
-            }
-            FileWriter fw = new FileWriter(dukeTxt, true);
-            BufferedWriter dukeWrite = new BufferedWriter(fw);
-            int i = tasks.size() - 1;
-            dukeWrite.write(tasks.get(i).toString());
-            dukeWrite.newLine();
-            dukeWrite.close();
-        } catch (IOException e) {
-            System.out.println("Oh no!!");
-        }
-    }
-
+    /***
+     * A method to delete all the contents in the arraylist and the txt file
+     * @param tasks an arraylist of all the tasks
+     * @throws IOException when the file cannot be found or created
+     */
     public void deleteAll(ArrayList<Task> tasks) throws IOException {
         tasks.clear();
         File dukeTxt = new File("duke.txt");
