+++ src/main/java/duke/ToDo.java
@@ -1,13 +1,23 @@
 package duke;
 
+/***
+ * A subclass of task indicating the type of task
+ */
 class ToDo extends Task {
     String icon = "[T]";
 
+    /***
+     * @param details details of the task
+     */
     public ToDo(String details) {
         super(details);
 
     }
 
+    /***
+     *
+     * @return a String type indicating the type of the task
+     */
     @Override
     public String toString() {
         return icon + super.toString();
